random_ops bug fix: write to stdout from child
Having the parent process read from the pipe and copy operations to
standard output resulted in standard output writes happening after
their corresponding database writes in cases where printing to standard
output should have come first. This resulted in a delete being written
to the database before the delete operation was written to standard
output, so when the operation history was cut off between the two, it
appeared that the key was lost spontaneously. Writing to both standard
output and the pipe to the parent simultaneously in the child process
removes this delay, and fixes the ordering of I/O operations.